Improve organisation in lib.rs
